backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
cryptography==43.0.1
httpx==0.27.2
anthropic==0.40.0
google-generativeai==0.8.3
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
